api/requirements.txt
fastapi==0.104.1
python-multipart==0.0.6
tensorflow-cpu==2.15.0
numpy==1.24.3
Pillow==10.1.0
pydantic==2.5.2
python-dotenv==1.0.0
mangum==0.17.0 